==> R_analysis/03_main_analyses/1_models_cancer_adjustments.Rmd <==
---
title: 'Modelling cancer, using different modelling strategies (overall and by sex)'
output:
  html_document:
    code_folding: hide
    highlight: tango
    theme: united
    toc: yes
    toc_depth: 4
    toc_float: yes
editor_options:
  chunk_output_type: console
---

```{r options, include=FALSE}
knitr::opts_chunk$set(
  echo = TRUE, warning = FALSE, message = FALSE, comment = FALSE,
  cache.lazy = FALSE
)
options(scipen = 999)
options(knitr.kable.NA = "")
```

```{r}
library(dplyr)
library(rms)
library(stringr)
library(tidyr)
library(kableExtra)
library(RColorBrewer)
library(forcats)
library(cowplot)
```


```{r}
load("M:/Elena - R/Multi-state  cancer/1 data prep/working.data.Rda")

# modify database for all transitions
e <- eapply(.GlobalEnv, function(x) {
  x <- x %>%
    mutate(smoke.all_time = factor(smoke.all_time)) %>%
    mutate(smoke.all_time = relevel(smoke.all_time, "Never smoker")) %>%
    mutate(medea = factor(medea)) %>%
    mutate(medea = relevel(medea, "U1")) %>%
    droplevels()
  output <- x
})


nms <- names(e)

for (n in nms) {
  assign(n, e[[n]])
}

# function for writing numbers
nice.num <- function(x) {
  prettyNum(x, big.mark = ",", nsmall = 0, digits = 0, scientific = FALSE)
}
nice.num2 <- function(x) {
  prettyNum(x, big.mark = ",", nsmall = 2, digits = 2, scientific = FALSE)
}
```

```{r}
# funcion for models
dd <<- datadist(r); options(datadist = "dd")

get.models <- function(transition.data,
                       transition.name) {
  # models<-list()
  summary.models <- NULL

  ## OVERALL MODELS + diff adjustments
  # 1. Unadjusted
  m.working <- cph(Surv(time, status) ~ cancer,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data
  )

  working.summary <- as.data.frame(summary(m.working, antilog = FALSE))
  working.summary$var <- row.names(working.summary)
  working.summary$gender <- "Overall"
  working.summary$c_status <- "Overall"
  working.summary$model <- "unadjusted"
  working.summary$transition.name <- transition.name
  working.summary <- working.summary %>%
    mutate(
      hr = exp(Effect),
      hr.low = exp(`Lower 0.95`),
      hr.high = exp(`Upper 0.95`)
    ) %>%
    select(transition.name, var, model, gender, c_status, hr, hr.low, hr.high)
  # output
  summary.models <- rbind(
    summary.models,
    working.summary
  )

  # 2 adjusted for age and gender
  # age based on previous analysis
  if (transition.name %in%
    c(
      "From general population to diagnosed with COVID-19",
      "From diagnosed with COVID-19 to hospitalised with COVID-19",
      "From general population to death"
    )) {
    m.working <- cph(Surv(time, status) ~ cancer
    + gender + rcs(age, 5),
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data
    )
  } else if
  (transition.name %in% c(
      "From general population to hospitalised with COVID-19",
      "From diagnosed with COVID-19 to death"
    )) {
    m.working <- cph(Surv(time, status) ~ cancer
    + gender + pol(age, 2),
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data, iter.max = 200
    )
  } else {
    m.working <- cph(Surv(time, status) ~ cancer
    + gender + rcs(age, 3),
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data
    )
  }
  working.summary <- as.data.frame(summary(m.working, antilog = FALSE))
  working.summary$var <- row.names(working.summary)
  working.summary$gender <- "Overall"
  working.summary$c_status <- "Overall"
  working.summary$model <- "Adjusted for age and sex"
  working.summary$transition.name <- transition.name
  working.summary <- working.summary %>%
    mutate(
      hr = exp(Effect),
      hr.low = exp(`Lower 0.95`),
      hr.high = exp(`Upper 0.95`)
    ) %>%
    select(transition.name, var, model, gender, c_status, hr, hr.low, hr.high)
  # output
  summary.models <- rbind(
    summary.models,
    working.summary
  )

  # 3 adjusted for age and gender, smoking and medea
  # age based on previous analysis
  if (transition.name %in%
    c(
      "From general population to diagnosed with COVID-19",
      "From general population to death",
      "From diagnosed with COVID-19 to hospitalised with COVID-19"
    )) {
    m.working <- cph(Surv(time, status) ~ cancer
    + gender + rcs(age, 5) +
      medea + smoke.all_time,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data
    )
  } else if
  (transition.name %in% c(
      "From general population to hospitalised with COVID-19",
      "From diagnosed with COVID-19 to death"
    )) {
    m.working <- cph(Surv(time, status) ~ cancer
    + gender + pol(age, 2) +
      medea + smoke.all_time,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data
    )
  } else {
    m.working <- cph(Surv(time, status) ~ cancer
    + gender + rcs(age, 3) +
      medea + smoke.all_time,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data
    )
  }
  working.summary <- as.data.frame(summary(m.working, medea = "U1", antilog = FALSE))
  working.summary$var <- row.names(working.summary)
  working.summary$gender <- "Overall"
  working.summary$c_status <- "Overall"
  working.summary$model <- "Mid adjustment"
  working.summary$transition.name <- transition.name
  working.summary <- working.summary %>%
    mutate(
      hr = exp(Effect),
      hr.low = exp(`Lower 0.95`),
      hr.high = exp(`Upper 0.95`)
    ) %>%
    select(transition.name, var, model, gender, c_status, hr, hr.low, hr.high)
  # output
  summary.models <- rbind(
    summary.models,
    working.summary
  )


  # 4 adjusted for age and gender, smoking and medea and comorbidities
  # age based on previous analysis

  if (transition.name %in%
    c(
      "From general population to diagnosed with COVID-19",
      "From general population to death",
      "From diagnosed with COVID-19 to hospitalised with COVID-19"
    )) {
    m.working <- cph(Surv(time, status) ~ cancer
    + gender + rcs(age, 5) +
      medea + smoke.all_time +
      a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
      a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data
    )
  } else if
  (transition.name %in% c(
      "From general population to hospitalised with COVID-19",
      "From diagnosed with COVID-19 to death"
    )) {
    m.working <- cph(Surv(time, status) ~ cancer
    + gender + pol(age, 2) +
      medea + smoke.all_time +
      a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
      a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data, iter.max = 200
    )
  } else {
    m.working <- cph(Surv(time, status) ~ cancer
    + gender + rcs(age, 3) +
      medea + smoke.all_time +
      a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
      a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data
    )
  }
  working.summary <- as.data.frame(summary(m.working, medea = "U1", antilog = FALSE))
  working.summary$var <- row.names(working.summary)
  working.summary$gender <- "Overall"
  working.summary$c_status <- "Overall"
  working.summary$model <- "Complete adjustment"
  working.summary$transition.name <- transition.name
  working.summary <- working.summary %>%
    mutate(
      hr = exp(Effect),
      hr.low = exp(`Lower 0.95`),
      hr.high = exp(`Upper 0.95`)
    ) %>%
    select(transition.name, var, model, gender, c_status, hr, hr.low, hr.high)
  # output
  summary.models <- rbind(
    summary.models,
    working.summary
  )

  summary.models


  ### MODELS STRATIFIED BY SEX
  ## MALES
  # 1. Unadjusted
  m.working <- cph(Surv(time, status) ~ cancer,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>% filter(gender == "Male")
  )

  working.summary <- as.data.frame(summary(m.working, antilog = FALSE))
  working.summary$var <- row.names(working.summary)
  working.summary$gender <- "Male"
  working.summary$c_status <- "Overall"
  working.summary$model <- "unadjusted"
  working.summary$transition.name <- transition.name
  working.summary <- working.summary %>%
    mutate(
      hr = exp(Effect),
      hr.low = exp(`Lower 0.95`),
      hr.high = exp(`Upper 0.95`)
    ) %>%
    select(transition.name, var, model, gender, c_status, hr, hr.low, hr.high)
  # output
  summary.models <- rbind(
    summary.models,
    working.summary
  )

  # 2 adjusted for age
  # age based on previous analysis
  if (transition.name %in%
    c(
      "From general population to diagnosed with COVID-19",
      "From diagnosed with COVID-19 to hospitalised with COVID-19",
      "From general population to death"
    )) {
    m.working <- cph(Surv(time, status) ~ cancer
    + rcs(age, 5),
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>% filter(gender == "Male")
    )
  } else if
  (transition.name %in% c(
      "From general population to hospitalised with COVID-19",
      "From diagnosed with COVID-19 to death"
    )) {
    m.working <- cph(Surv(time, status) ~ cancer
    + pol(age, 2),
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>% filter(gender == "Male"), iter.max = 200
    )
  } else {
    m.working <- cph(Surv(time, status) ~ cancer
    + rcs(age, 3),
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>% filter(gender == "Male"), iter.max = 200
    )
  }
  working.summary <- as.data.frame(summary(m.working, antilog = FALSE))
  working.summary$var <- row.names(working.summary)
  working.summary$gender <- "Male"
  working.summary$c_status <- "Overall"
  working.summary$model <- "Adjusted for age and sex"
  working.summary$transition.name <- transition.name
  working.summary <- working.summary %>%
    mutate(
      hr = exp(Effect),
      hr.low = exp(`Lower 0.95`),
      hr.high = exp(`Upper 0.95`)
    ) %>%
    select(transition.name, var, model, gender, c_status, hr, hr.low, hr.high)
  # output
  summary.models <- rbind(
    summary.models,
    working.summary
  )

  # 3 adjusted for age, smoking and medea
  # age based on previous analysis
  if (transition.name %in%
    c(
      "From general population to diagnosed with COVID-19",
      "From general population to death",
      "From diagnosed with COVID-19 to hospitalised with COVID-19"
    )) {
    m.working <- cph(Surv(time, status) ~ cancer
    + rcs(age, 5) +
      medea + smoke.all_time,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>% filter(gender == "Male"), iter.max = 200
    )
  } else if
  (transition.name %in% c(
      "From general population to hospitalised with COVID-19",
      "From diagnosed with COVID-19 to death"
    )) {
    m.working <- cph(Surv(time, status) ~ cancer
    + pol(age, 2) +
      medea + smoke.all_time,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>% filter(gender == "Male"), iter.max = 200
    )
  } else {
    m.working <- cph(Surv(time, status) ~ cancer
    + rcs(age, 3) +
      medea + smoke.all_time,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>% filter(gender == "Male"), iter.max = 200
    )
  }
  working.summary <- as.data.frame(summary(m.working, medea = "U1", antilog = FALSE))
  working.summary$var <- row.names(working.summary)
  working.summary$gender <- "Male"
  working.summary$c_status <- "Overall"
  working.summary$model <- "Mid adjustment"
  working.summary$transition.name <- transition.name
  working.summary <- working.summary %>%
    mutate(
      hr = exp(Effect),
      hr.low = exp(`Lower 0.95`),
      hr.high = exp(`Upper 0.95`)
    ) %>%
    select(transition.name, var, model, gender, c_status, hr, hr.low, hr.high)
  # output
  summary.models <- rbind(
    summary.models,
    working.summary
  )


  # 4 adjusted for age, smoking and medea and comorbidities
  # age based on previous analysis

  if (transition.name %in%
    c(
      "From general population to diagnosed with COVID-19",
      "From general population to death",
      "From diagnosed with COVID-19 to hospitalised with COVID-19"
    )) {
    m.working <- cph(Surv(time, status) ~ cancer
    + rcs(age, 5) +
      medea + smoke.all_time +
      a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
      a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>% filter(gender == "Male"), iter.max = 200
    )
  } else if
  (transition.name %in% c(
      "From general population to hospitalised with COVID-19",
      "From diagnosed with COVID-19 to death"
    )) {
    m.working <- cph(Surv(time, status) ~ cancer
    + pol(age, 2) +
      medea + smoke.all_time +
      a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
      a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>% filter(gender == "Male"), iter.max = 200
    )
  } else {
    m.working <- cph(Surv(time, status) ~ cancer
    + rcs(age, 3) +
      medea + smoke.all_time +
      a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
      a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>% filter(gender == "Male"), iter.max = 200
    )
  }
  working.summary <- as.data.frame(summary(m.working, medea = "U1", antilog = FALSE))
  working.summary$var <- row.names(working.summary)
  working.summary$gender <- "Male"
  working.summary$c_status <- "Overall"
  working.summary$model <- "Complete adjustment"
  working.summary$transition.name <- transition.name
  working.summary <- working.summary %>%
    mutate(
      hr = exp(Effect),
      hr.low = exp(`Lower 0.95`),
      hr.high = exp(`Upper 0.95`)
    ) %>%
    select(transition.name, var, model, gender, c_status, hr, hr.low, hr.high)
  # output
  summary.models <- rbind(
    summary.models,
    working.summary
  )

  summary.models

  ## FEMALES
  # 1. Unadjusted -----
  m.working <- cph(Surv(time, status) ~ cancer,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>% filter(gender == "Female")
  )

  working.summary <- as.data.frame(summary(m.working, antilog = FALSE))
  working.summary$var <- row.names(working.summary)
  working.summary$gender <- "Female"
  working.summary$c_status <- "Overall"
  working.summary$model <- "unadjusted"
  working.summary$transition.name <- transition.name
  working.summary <- working.summary %>%
    mutate(
      hr = exp(Effect),
      hr.low = exp(`Lower 0.95`),
      hr.high = exp(`Upper 0.95`)
    ) %>%
    select(transition.name, var, model, gender, c_status, hr, hr.low, hr.high)
  # output
  summary.models <- rbind(
    summary.models,
    working.summary
  )

  # 2 adjusted for age
  # age based on previous analysis
  if (transition.name %in%
    c(
      "From general population to diagnosed with COVID-19",
      "From diagnosed with COVID-19 to hospitalised with COVID-19",
      "From general population to death"
    )) {
    m.working <- cph(Surv(time, status) ~ cancer
    + rcs(age, 5),
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>% filter(gender == "Female")
    )
  } else if
  (transition.name %in% c(
      "From general population to hospitalised with COVID-19",
      "From diagnosed with COVID-19 to death"
    )) {
    m.working <- cph(Surv(time, status) ~ cancer
    + pol(age, 2),
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>% filter(gender == "Female"), iter.max = 200
    )
  } else {
    m.working <- cph(Surv(time, status) ~ cancer
    + rcs(age,3),
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>% filter(gender == "Female"), iter.max = 200
    )
  }
  working.summary <- as.data.frame(summary(m.working, antilog = FALSE))
  working.summary$var <- row.names(working.summary)
  working.summary$gender <- "Female"
  working.summary$c_status <- "Overall"
  working.summary$model <- "Adjusted for age and sex"
  working.summary$transition.name <- transition.name
  working.summary <- working.summary %>%
    mutate(
      hr = exp(Effect),
      hr.low = exp(`Lower 0.95`),
      hr.high = exp(`Upper 0.95`)
    ) %>%
    select(transition.name, var, model, gender, c_status, hr, hr.low, hr.high)
  # output
  summary.models <- rbind(
    summary.models,
    working.summary
  )

  # 3 adjusted for age, smoking and medea
  # age based on previous analysis
  if (transition.name %in%
    c(
      "From general population to diagnosed with COVID-19",
      "From general population to death",
      "From diagnosed with COVID-19 to hospitalised with COVID-19"
    )) {
    m.working <- cph(Surv(time, status) ~ cancer
    + rcs(age, 5) +
      medea + smoke.all_time,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>% filter(gender == "Female"), iter.max = 200
    )
  } else if
  (transition.name %in% c(
      "From general population to hospitalised with COVID-19",
      "From diagnosed with COVID-19 to death"
    )) {
    m.working <- cph(Surv(time, status) ~ cancer
    + pol(age, 2) +
      medea + smoke.all_time,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>% filter(gender == "Female"), iter.max = 200
    )
  } else {
    m.working <- cph(Surv(time, status) ~ cancer
    + rcs(age, 3) +
      medea + smoke.all_time,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>% filter(gender == "Female"), iter.max = 200
    )
  }
  working.summary <- as.data.frame(summary(m.working, medea = "U1", antilog = FALSE))
  working.summary$var <- row.names(working.summary)
  working.summary$gender <- "Female"
  working.summary$c_status <- "Overall"
  working.summary$model <- "Mid adjustment"
  working.summary$transition.name <- transition.name
  working.summary <- working.summary %>%
    mutate(
      hr = exp(Effect),
      hr.low = exp(`Lower 0.95`),
      hr.high = exp(`Upper 0.95`)
    ) %>%
    select(transition.name, var, model, gender, c_status, hr, hr.low, hr.high)
  # output
  summary.models <- rbind(
    summary.models,
    working.summary
  )



  # 4 adjusted for age, smoking and medea and comorbidities
  # age based on previous analysis

  if (transition.name %in%
    c(
      "From general population to diagnosed with COVID-19",
      "From general population to death",
      "From diagnosed with COVID-19 to hospitalised with COVID-19"
    )) {
    m.working <- cph(Surv(time, status) ~ cancer
    + rcs(age, 5) +
      medea + smoke.all_time +
      a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
      a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>% filter(gender == "Female"), iter.max = 200
    )
  } else if
  (transition.name %in% c(
      "From general population to hospitalised with COVID-19",
      "From diagnosed with COVID-19 to death"
    )) {
    m.working <- cph(Surv(time, status) ~ cancer
    + pol(age, 2) +
      medea + smoke.all_time +
      a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
      a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>% filter(gender == "Female"), iter.max = 200
    )
  } else {
    m.working <- cph(Surv(time, status) ~ cancer
    + rcs(age, 3) +
      medea + smoke.all_time +
      a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
      a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>% filter(gender == "Female"), iter.max = 200
    )
  }
  working.summary <- as.data.frame(summary(m.working, medea = "U1", antilog = FALSE))
  working.summary$var <- row.names(working.summary)
  working.summary$gender <- "Female"
  working.summary$c_status <- "Overall"
  working.summary$model <- "Complete adjustment"
  working.summary$transition.name <- transition.name
  working.summary <- working.summary %>%
    mutate(
      hr = exp(Effect),
      hr.low = exp(`Lower 0.95`),
      hr.high = exp(`Upper 0.95`)
    ) %>%
    select(transition.name, var, model, gender, c_status, hr, hr.low, hr.high)
  # output
  summary.models <- rbind(
    summary.models,
    working.summary
  )

  summary.models
}
```


```{r}
# get results
summary.m.healthy.diagnosis <- get.models(
  transition.data = r.healthy.diagnosis,
  transition.name = "From general population to diagnosed with COVID-19"
)

summary.m.healthy.hospitalised <- get.models(
  transition.data = r.healthy.hospitalised,
  transition.name =  "From general population to hospitalised with COVID-19"
)

summary.m.healthy.death <- get.models(
  transition.data = r.healthy.death,
  transition.name = "From general population to death"
)

summary.m.diagnosis.hospitalised <- get.models(
  transition.data = r.diagnosis.hospitalised,
  transition.name = "From diagnosed with COVID-19 to hospitalised with COVID-19"
)

summary.m.diagnosis.death <- get.models(
  transition.data = r.diagnosis.death,
  transition.name = "From diagnosed with COVID-19 to death"
)

summary.m.hospitalised.death <- get.models(
  transition.data = r.hospitalised.death,
  transition.name = "From hospitalised with COVID-19 to death"
)

estimates <- rbind(
  summary.m.healthy.diagnosis,
  summary.m.healthy.hospitalised,
  summary.m.healthy.death,
  summary.m.diagnosis.hospitalised,
  summary.m.diagnosis.death,
  summary.m.hospitalised.death
)

# save results
# save(estimates, file = "M:/Elena - R/Multi-state  cancer/results/1.estimates_adjustments_gender.Rda")
# load("M:/Elena - R/Multi-state  cancer/results/1.estimates_adjustments_gender.Rda")
```



```{r}
# Results table - we only keep HR for our variable of interest (cancer)
summary.table <- estimates %>%
  mutate(est = paste0(
    nice.num2(hr), " (",
    nice.num2(hr.low), " to ",
    nice.num2(hr.high), ")"
  )) %>%
  filter(var == "cancer") %>%
  select(transition.name, model, gender, est)

summary.table <- summary.table %>%
  pivot_wider(
    names_from = model,
    values_from = est
  )

# write.csv2(summary.table, "M:/Elena - R/Multi-state  cancer/results/1.HR_cancer_models_gender.csv")
```

# HR for patients with cancer compared to cancer-free patients, overall and by sex. Unadjusted and adjusted models 
```{r}
kable(summary.table,
  col.names = c(
    "Transition", "Sex",
    "Unadjusted", "Age and sex", "Age, sex,\nsmoking and MEDEA", "Age, sex,\nsmoking, MEDEA \nand comorbidities"
  )
) %>%
  kable_styling(bootstrap_options = c("striped", "bordered")) %>%
  collapse_rows(columns = 1)
```

# HR for patients with cancer compared to cancer-free patients, overall and by sex. Fully adjusted models
```{r}

summary.table2 <- summary.table %>%
  select(1, 2, 6)

summary.table2 <- summary.table2 %>%
  pivot_wider(
    names_from = 2,
    values_from = 3
  )

kable(summary.table2,
  col.names = c("Transition", "Overall", "Male", "Female")
) %>%
  kable_styling(bootstrap_options = c("striped", "bordered")) %>%
  collapse_rows(columns = 1)
```


#Figures
```{r}
#  the data
plot.data <- estimates %>%
  mutate(
    transition.name =
      ifelse(transition.name ==
        "From general population to diagnosed with COVID-19",
      "From general population\nto diagnosed\nwith COVID-19",
      ifelse(transition.name ==
        "From general population to hospitalised with COVID-19",
      "From general population\nto hospitalised\nwith COVID-19",
      ifelse(transition.name ==
        "From diagnosed with COVID-19 to hospitalised with COVID-19",
      "From diagnosed with\nCOVID-19 to hospitalised\nwith COVID-19",
      ifelse(transition.name ==
        "From diagnosed with COVID-19 to death",
      "From diagnosed with\nCOVID-19 to\ndeath",
      ifelse(transition.name ==
        "From hospitalised with COVID-19 to death",
      "From hospitalised\nwith COVID-19\nto death",
      ifelse(transition.name ==
        "From general population to death",
      "From general population\nto death",
      NA
      )
      )
      )
      )
      )
      )
  ) %>%
  mutate(transition.name = factor(transition.name,
    levels = c(
      "From general population\nto diagnosed\nwith COVID-19",
      "From general population\nto hospitalised\nwith COVID-19",
      "From diagnosed with\nCOVID-19 to hospitalised\nwith COVID-19",
      "From diagnosed with\nCOVID-19 to\ndeath",
      "From hospitalised\nwith COVID-19\nto death",
      "From general population\nto death"
    )
  ))

plot.data <- plot.data %>%
  mutate(hr_text = paste0(
    nice.num2(hr), " (",
    nice.num2(hr.low), " to ",
    nice.num2(hr.high), ")"
  )) %>%
  mutate(gender = factor(gender, levels = c("Male", "Female", "Overall"))) %>%
  mutate(model = ifelse(model == "unadjusted", "Unadjusted",
    ifelse(model == "Adjusted for age and sex", "Adjusted for age and sex",
      ifelse(model == "Mid adjustment", "Adjusted for age, sex, MEDEA and smoking",
        ifelse(model == "Complete adjustment",
          "Adjusted for age, sex, MEDEA, smoking and\ncomorbidities (main models)",
          NA
        )
      )
    )
  )) %>%
  mutate(model = factor(model,
    levels = c(
      "Adjusted for age, sex, MEDEA, smoking and\ncomorbidities (main models)",
      "Adjusted for age, sex, MEDEA and smoking",
      "Adjusted for age and sex",
      "Unadjusted"
    )
  ))

plot.data <- plot.data %>%
  filter(var == "cancer")

# Specify font
windowsFonts(A = windowsFont("Times New Roman"))
```

# Figures
## HR for risk of COVID-19 outcomes in patients with cancer relative to cancer-free patients, using different modelling strategies
```{r}
# without from general population to death
plot.data2 <- plot.data %>%
  filter(gender == "Overall")

hrs.all.models <- plot.data2 %>%
  filter(transition.name != "From general population\nto death") %>%
  filter(model!= "Unadjusted") %>%
  ggplot(aes(
    x = hr,
    y = model,
    xmin = hr.low,
    xmax = hr.high,
    color = model
  )) +
  scale_x_log10(
    breaks = c(0.8, 1, 2),
    labels = c( "0.8", "1", "2"),
    limits = c(0.7, 2.2)
  ) +
  facet_grid(transition.name ~ .,
    drop = TRUE,
    #    scales = "free_y",
    #   space = "free_y",
    switch = "y"
  ) +
  geom_point(size = 3, position = position_dodge(width = 1), shape = 20) +
  geom_errorbar(width = 0, size = 1, position = position_dodge(width = 1)) +
  scale_colour_manual(values = c("#2660A4", "#1A936F", "#FAA916", "#1B1B1E")) +
  geom_vline(
    xintercept = 1, colour = "#000000",
    linetype = 2
  ) +
  theme_bw() +
  theme(
    text = element_text(family = "A"),
    panel.spacing = unit(0, "lines"),
    legend.position = "bottom",
    legend.title = element_blank(),
    #  legend.position=c(1,1),
    legend.justification = c("left"),
    legend.text = element_text(size = 14),
    axis.title.y = element_blank(),
    axis.text.x = element_text(size = 14),
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    axis.title = element_text(size = 14),
    strip.text.x = element_text(size = 12, face = "bold"),
    strip.text.y.left = element_text(size = 14, angle = 0, face = "bold"),
    strip.background = element_rect(fill = "#f7f7f7")
  ) +
  xlab("\n HR (with 95% CI)") +
  ylab("") +
  guides(colour = guide_legend(reverse = T, nrow = 4))


data_table <- plot.data2 %>%
  filter(transition.name != "From general population\nto death") %>%
    filter(model!= "Unadjusted") %>%
  ggplot(aes(x = hr, y = model)) +
  facet_grid(transition.name ~ .,
    drop = TRUE, switch = "y"
  ) +
  scale_x_log10(limits = c(0.1, 0.6)) +
  geom_text(aes(x = 0.2, y = model, label = hr_text),
    position = position_dodge(width = 1),
    family = "A",
    size = 5
  ) +
  theme(
    text = element_text(family = "A"),
    axis.ticks.y = element_blank(),
    axis.ticks.x = element_blank(),
    legend.position = "none",
    legend.title = element_blank(),
    panel.grid.minor = element_blank(),
    panel.grid.major = element_blank(),
    axis.title.y = element_blank(),
    axis.text = element_blank(),
    panel.background = element_blank(),
    strip.background = element_blank(),
    strip.text.y.left = element_blank(),
    plot.title = element_text(hjust = 0.3, size = 13, face = "bold")
  ) +
  labs(title = " HR (95% CI)") +
  xlab("")

plot <- plot_grid(hrs.all.models, data_table, align = "h", axis = "bt", rel_widths = c(1, 0.35))
plot


# ggsave("M:/Elena - R/Multi-state  cancer/results/gg.hrs.overall.models_cov.png", plot, dpi = 400,
  # width = 9, height = 9 )
```


## HR for risk of COVID-19 outcomes in patients with cancer relative to cancer-free patients, including unadjusted models 

```{r}


hrs.all.models <- plot.data2 %>%
  ggplot(aes(
    x = hr,
    y = model,
    xmin = hr.low,
    xmax = hr.high,
    color = model
  )) +
  scale_x_log10(
    breaks = c(0.5, 0.75, 1, 2, 5, 10),
    labels = c("0.5", "", "1", "2", "5", "10"),
    limits = c(0.5, 10)
  ) +
  facet_grid(transition.name ~ .,
    drop = TRUE,
    switch = "y"
  ) +
  geom_point(size = 3, position = position_dodge(width = 1), shape = 20) +
  geom_errorbar(width = 0, size = 1, position = position_dodge(width = 1)) +
  scale_colour_manual(values = c("#2660A4", "#1A936F", "#FAA916", "#1B1B1E")) +
  geom_vline(
    xintercept = 1, colour = "#000000",
    linetype = 2
  ) +
  theme_bw() +
  theme(
    text = element_text(family = "A"),
    panel.spacing = unit(0, "lines"),
    legend.position = "bottom",
    legend.title = element_blank(),
    legend.justification = c("left"),
    legend.text = element_text(size = 14),
    axis.title.y = element_blank(),
    axis.text.x = element_text(size = 14),
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    axis.title = element_text(size = 14),
    strip.text.x = element_text(size = 12, face = "bold"),
    strip.text.y.left = element_text(size = 14, angle = 0, face = "bold"),
    strip.background = element_rect(fill = "#f7f7f7")
  ) +
  xlab("\n aHR (with 95% CI)") +
  ylab("") +
  guides(colour = guide_legend(reverse = T, nrow = 4))


data_table <- ggplot(
  data = plot.data2,
  aes(x = hr, y = model)
) +
  facet_grid(transition.name ~ .,
    drop = TRUE, switch = "y"
  ) +
  geom_text(aes(x = 0, y = model, label = hr_text),
    position = position_dodge(width = 1),
    family = "A",
    size = 5
  ) +
  theme(
    text = element_text(family = "A"),
    axis.ticks.y = element_blank(),
    axis.ticks.x = element_blank(),
    legend.position = "none",
    legend.title = element_blank(),
    panel.grid.minor = element_blank(),
    panel.grid.major = element_blank(),
    axis.title.y = element_blank(),
    axis.text = element_blank(),
    panel.background = element_blank(),
    strip.background = element_blank(),
    strip.text.y.left = element_blank(),
    plot.title = element_text(hjust = 0.5, size = 13, face = "bold")
  ) +
  labs(title = " HR (with 95% CI)") +
  xlab("")



plot2 <- plot_grid(hrs.all.models, data_table, align = "h", axis = "bt", rel_widths = c(1, 0.3))

plot2


# ggsave("M:/Elena - R/Multi-state  cancer/results/gg.hrs.overall.models.png", plot2, dpi = 400,
#  width = 8, height = 9)
```
