==> R_analysis/03_main_analyses/4_models_years_cancer_solid_cancers.Rmd <==
---
title: 'Modelling solid cancer types, overall and stratified by years since cancer diagnosis'
output:
  html_document:
    code_folding: hide
    highlight: tango
    theme: united
    toc: yes
    toc_depth: 4
    toc_float: yes
highlight: tango
theme: united
toc: yes
toc_depth: 4
toc_float: yes
editor_options:
  chunk_output_type: console
---
  
```{r options, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE, comment = FALSE)
options(scipen = 999)
```

```{r packages, include=FALSE}
# packages -----
# install.packages("seplyr")
library(kableExtra)
library(dplyr)
library(ggplot2)
library(rms)
library(stargazer)
library(knitr)
library(stringr)
library(tidyr)
library(seplyr)

# function for numbers
nice.num <- function(x) {
  prettyNum(x, big.mark = ",", nsmall = 0, digits = 0, scientific = FALSE)
}
nice.num2 <- function(x) {
  prettyNum(x, big.mark = ",", nsmall = 2, digits = 2, scientific = FALSE)
}

# Font for plots
windowsFonts(A = windowsFont("Times New Roman"))
```


```{r data, include=FALSE}
load("M:/Elena - R/Multi-state  cancer/1 data prep/working.data.Rda")

dd <<- datadist(r); options(datadist = "dd")

```


```{r}
# Cancers to analyse
cancers <- c(
  "colorectal",
  "bladder",
  "lung_trachea"
#  "kidney",
#  "melanoma"
)
```


```{r}
# function for models - only fully adjusted models
get.models <- function(cancers, # var of interest - #all cancers, breast, and prostate separately
                       transition.data, # data for transition of interest
                       transition.name) {

  # models<-list()
  summary.models <- NULL

  # Breast - only with female
  # age modeling based on prior results
  ## Overall
  if (transition.name %in%
    c(
      "From general population to diagnosed with COVID-19",
      "From diagnosed with COVID-19 to hospitalised with COVID-19",
      "From general population to death"
    )) {
    m.working <- cph(Surv(time, status) ~ breast +
      rcs(age, 5) +
      medea + smoke.all_time +
      a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
      a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>%
      filter(gender == "Female" & (cancer == 0 | breast == 1)),
    iter.max = 200
    )
  } else if (transition.name %in%
    c(
      "From general population to hospitalised with COVID-19",
      "From diagnosed with COVID-19 to death"
    )) {
    m.working <- cph(Surv(time, status) ~ breast +
      pol(age, 2) +
      medea + smoke.all_time +
      a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
      a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>%
      filter(gender == "Female" & (cancer == 0 | breast == 1)),
    iter.max = 200
    )
  } else {
    m.working <- cph(Surv(time, status) ~ breast +
      rcs(age, 3) +
      medea + smoke.all_time +
      a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
      a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>%
      filter(gender == "Female" & (cancer == 0 | breast == 1)),
    iter.max = 200
    )
  }
  working.summary <- as.data.frame(summary(m.working, antilog = FALSE))
  working.summary$var <- row.names(working.summary)
  working.summary$gender <- "Overall"
  working.summary$c_status <- "Overall"
  working.summary$model <- "Fully adjusted"
  working.summary$transition.name <- transition.name
  working.summary <- working.summary %>%
    mutate(
      hr = exp(Effect),
      hr.low = exp(`Lower 0.95`),
      hr.high = exp(`Upper 0.95`)
    ) %>%
    select(
      transition.name, var, model, gender,
      c_status, hr, hr.low, hr.high
    )
  # output
  summary.models <- rbind(
    summary.models,
    working.summary
  )


  ## Breast cancers diagnosed <5 years - # we use the varaible c_status_2cat 
  # (1-5 and <1 year together due to small sample sizes)
  if (transition.name %in%
    c(
      "From general population to diagnosed with COVID-19",
      "From diagnosed with COVID-19 to hospitalised with COVID-19",
      "From general population to death"
    )) {
    m.working <- cph(Surv(time, status) ~ breast +
      rcs(age, 5) +
      medea + smoke.all_time +
      a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
      a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>%
      filter(gender == "Female" &
        (cancer == 0 | (breast == 1 & c_status_2cat == "<5 years"))),
    iter.max = 200
    )
  } else if (transition.name %in%
    c(
      "From general population to hospitalised with COVID-19",
      "From diagnosed with COVID-19 to death"
    )) {
    m.working <- cph(Surv(time, status) ~ breast +
      pol(age, 2) +
      medea + smoke.all_time +
      a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
      a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>%
      filter(gender == "Female" &
        (cancer == 0 | (breast == 1 & c_status_2cat == "<5 years"))),
    iter.max = 200
    )
  } else {
    m.working <- cph(Surv(time, status) ~ breast +
      rcs(age, 3) +
      medea + smoke.all_time +
      a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
      a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>%
      filter(gender == "Female" &
        (cancer == 0 | (breast == 1 & c_status_2cat == "<5 years"))),
    iter.max = 200
    )
  }
  working.summary <- as.data.frame(summary(m.working, antilog = FALSE))
  working.summary$var <- row.names(working.summary)
  working.summary$gender <- "Overall"
  working.summary$c_status <- "<5 years"
  working.summary$model <- "Fully adjusted"
  working.summary$transition.name <- transition.name
  working.summary <- working.summary %>%
    mutate(
      hr = exp(Effect),
      hr.low = exp(`Lower 0.95`),
      hr.high = exp(`Upper 0.95`)
    ) %>%
    select(
      transition.name, var, model, gender,
      c_status, hr, hr.low, hr.high
    )
  # output
  summary.models <- rbind(
    summary.models,
    working.summary
  )

  ## breast cancers diagnosed >5y ago
  if (transition.name %in%
    c(
      "From general population to diagnosed with COVID-19",
      "From diagnosed with COVID-19 to hospitalised with COVID-19",
      "From general population to death"
    )) {
    m.working <- cph(Surv(time, status) ~ breast +
      rcs(age, 5) +
      medea + smoke.all_time +
      a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
      a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>%
      filter(gender == "Female" &
        (cancer == 0 | (breast == 1 & c_status_2cat == ">5years"))),
    iter.max = 200
    )
  } else if (transition.name %in%
    c(
      "From general population to hospitalised with COVID-19",
      "From diagnosed with COVID-19 to death"
    )) {
    m.working <- cph(Surv(time, status) ~ breast +
      pol(age, 2) +
      medea + smoke.all_time +
      a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
      a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>%
      filter(gender == "Female" &
        (cancer == 0 | (breast == 1 & c_status_2cat == ">5years"))),
    iter.max = 200
    )
  } else {
    m.working <- cph(Surv(time, status) ~ breast +
      rcs(age, 3) +
      medea + smoke.all_time +
      a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
      a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>%
      filter(gender == "Female" &
        (cancer == 0 | (breast == 1 & c_status_2cat == ">5years"))),
    iter.max = 200
    )
  }
  working.summary <- as.data.frame(summary(m.working, antilog = FALSE))
  working.summary$var <- row.names(working.summary)
  working.summary$gender <- "Overall"
  working.summary$c_status <- ">5years"
  working.summary$model <- "Fully adjusted"
  working.summary$transition.name <- transition.name
  working.summary <- working.summary %>%
    mutate(
      hr = exp(Effect),
      hr.low = exp(`Lower 0.95`),
      hr.high = exp(`Upper 0.95`)
    ) %>%
    select(
      transition.name, var, model, gender,
      c_status, hr, hr.low, hr.high
    )
  # output
  summary.models <- rbind(
    summary.models,
    working.summary
  )

  summary.models

  # prostate - only males
  ##  Overall
  if (transition.name %in%
    c(
      "From general population to diagnosed with COVID-19",
      "From diagnosed with COVID-19 to hospitalised with COVID-19",
      "From general population to death"
    )) {
    m.working <- cph(Surv(time, status) ~ prostate +
      rcs(age, 5) +
      medea + smoke.all_time +
      a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
      a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>%
      filter(gender == "Male" &
        (cancer == 0 | prostate == 1)),
    iter.max = 200
    )
  } else if (transition.name %in%
    c(
      "From general population to hospitalised with COVID-19",
      "From diagnosed with COVID-19 to death"
    )) {
    m.working <- cph(Surv(time, status) ~ prostate +
      pol(age, 2) +
      medea + smoke.all_time +
      a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
      a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>%
      filter(gender == "Male" &
        (cancer == 0 | prostate == 1)),
    iter.max = 200
    )
  } else {
    m.working <- cph(Surv(time, status) ~ prostate +
      rcs(age, 3) +
      medea + smoke.all_time +
      a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
      a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>%
      filter(gender == "Male" &
        (cancer == 0 | prostate == 1)),
    iter.max = 200
    )
  }
  working.summary <- as.data.frame(summary(m.working, antilog = FALSE))
  working.summary$var <- row.names(working.summary)
  working.summary$gender <- "Overall"
  working.summary$c_status <- "Overall"
  working.summary$model <- "Fully adjusted"
  working.summary$transition.name <- transition.name
  working.summary <- working.summary %>%
    mutate(
      hr = exp(Effect),
      hr.low = exp(`Lower 0.95`),
      hr.high = exp(`Upper 0.95`)
    ) %>%
    select(
      transition.name, var, model, gender,
      c_status, hr, hr.low, hr.high
    )
  # output
  summary.models <- rbind(
    summary.models,
    working.summary
  )



  # Prostate cancers diagnosed <5 years ago
  if (transition.name %in%
    c(
      "From general population to diagnosed with COVID-19",
      "From diagnosed with COVID-19 to hospitalised with COVID-19",
      "From general population to death"
    )) {
    m.working <- cph(Surv(time, status) ~ prostate +
      rcs(age, 5) +
      medea + smoke.all_time +
      a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
      a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>%
      filter(gender == "Male" &
        (cancer == 0 | (prostate == 1 & c_status_2cat == "<5 years"))),
    iter.max = 200
    )
  } else if (transition.name %in%
    c(
      "From general population to hospitalised with COVID-19",
      "From diagnosed with COVID-19 to death"
    )) {
    m.working <- cph(Surv(time, status) ~ prostate +
      pol(age, 2) +
      medea + smoke.all_time +
      a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
      a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>%
      filter(gender == "Male" &
        (cancer == 0 | (prostate == 1 & c_status_2cat == "<5 years"))),
    iter.max = 200
    )
  } else {
    m.working <- cph(Surv(time, status) ~ prostate +
      rcs(age, 3) +
      medea + smoke.all_time +
      a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
      a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>%
      filter(gender == "Male" &
        (cancer == 0 | (prostate == 1 & c_status_2cat == "<5 years"))),
    iter.max = 200
    )
  }
  working.summary <- as.data.frame(summary(m.working, antilog = FALSE))
  working.summary$var <- row.names(working.summary)
  working.summary$gender <- "Overall"
  working.summary$c_status <- "<5 years"
  working.summary$model <- "Fully adjusted"
  working.summary$transition.name <- transition.name
  working.summary <- working.summary %>%
    mutate(
      hr = exp(Effect),
      hr.low = exp(`Lower 0.95`),
      hr.high = exp(`Upper 0.95`)
    ) %>%
    select(
      transition.name, var, model, gender,
      c_status, hr, hr.low, hr.high
    )
  # output
  summary.models <- rbind(
    summary.models,
    working.summary
  )

  ## Prostate cancers diagnosed >5y ago
  if (transition.name %in%
    c(
      "From general population to diagnosed with COVID-19",
      "From diagnosed with COVID-19 to hospitalised with COVID-19",
      "From general population to death"
    )) {
    m.working <- cph(Surv(time, status) ~ prostate +
      rcs(age, 5) +
      medea + smoke.all_time +
      a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
      a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>%
      filter(gender == "Male" &
        (cancer == 0 | (prostate == 1 & c_status_2cat == ">5years"))),
    iter.max = 200
    )
  } else if (transition.name %in%
    c(
      "From general population to hospitalised with COVID-19",
      "From diagnosed with COVID-19 to death"
    )) {
    m.working <- cph(Surv(time, status) ~ prostate +
      pol(age, 2) +
      medea + smoke.all_time +
      a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
      a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>%
      filter(gender == "Male" &
        (cancer == 0 | (prostate == 1 & c_status_2cat == ">5years"))),
    iter.max = 200
    )
  } else {
    m.working <- cph(Surv(time, status) ~ prostate +
      rcs(age, 3) +
      medea + smoke.all_time +
      a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
      a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity,
    surv = TRUE, x = TRUE, y = TRUE,
    data = transition.data %>%
      filter(gender == "Male" &
        (cancer == 0 | (prostate == 1 & c_status_2cat == ">5years"))),
    iter.max = 200
    )
  }
  working.summary <- as.data.frame(summary(m.working, antilog = FALSE))
  working.summary$var <- row.names(working.summary)
  working.summary$gender <- "Overall"
  working.summary$c_status <- ">5years"
  working.summary$model <- "Fully adjusted"
  working.summary$transition.name <- transition.name
  working.summary <- working.summary %>%
    mutate(
      hr = exp(Effect),
      hr.low = exp(`Lower 0.95`),
      hr.high = exp(`Upper 0.95`)
    ) %>%
    select(
      transition.name, var, model, gender,
      c_status, hr, hr.low, hr.high
    )
  # output
  summary.models <- rbind(
    summary.models,
    working.summary
  )

  summary.models


  ## other cancers

  for (i in 1:length(cancers)) {
    message(paste0("working on cancer ", i, "( ", cancers[i], ")"))
    working.cancer <- cancers[i]
    # overall models -----

    #  Overall
    if (transition.name %in%
      c(
        "From general population to diagnosed with COVID-19",
        "From diagnosed with COVID-19 to hospitalised with COVID-19",
        "From general population to death"
      )) {
      m.working <- cph(as.formula(paste(
        "Surv(time, status)~ ",
        paste(
          cancers[i],
          "+gender+rcs(age,5)+
                medea+smoke.all_time+
                a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
                a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity"
        )
      )),
      surv = TRUE, x = TRUE, y = TRUE,
      data = transition.data %>%
        filter_se(paste0("cancer==0 | ", cancers[i], "==1")),
      iter.max = 200
      )
    } else if (transition.name %in%
      c(
        "From general population to hospitalised with COVID-19",
        "From diagnosed with COVID-19 to death"
      )) {
      m.working <- cph(as.formula(paste(
        "Surv(time, status)~ ",
        paste(
          cancers[i],
          "+gender+pol(age,2)+
                medea+smoke.all_time+
                a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
                a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity"
        )
      )),
      surv = TRUE, x = TRUE, y = TRUE,
      data = transition.data %>%
        filter_se(paste0("cancer==0 | ", cancers[i], "==1")),
      iter.max = 200
      )
    } else {
      m.working <- cph(as.formula(paste(
        "Surv(time, status)~ ",
        paste(
          cancers[i],
          "+gender+rcs(age,3)+
                medea+smoke.all_time+
                a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
                a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity"
        )
      )),
      surv = TRUE, x = TRUE, y = TRUE,
      data = transition.data %>%
        filter_se(paste0("cancer==0 | ", cancers[i], "==1")),
      iter.max = 200
      )
    }
    working.summary <- as.data.frame(summary(m.working, antilog = FALSE))
    working.summary$var <- row.names(working.summary)
    working.summary$gender <- "Overall"
    working.summary$c_status <- "Overall"
    working.summary$model <- "Fully adjusted"
    working.summary$transition.name <- transition.name
    working.summary <- working.summary %>%
      mutate(
        hr = exp(Effect),
        hr.low = exp(`Lower 0.95`),
        hr.high = exp(`Upper 0.95`)
      ) %>%
      select(
        transition.name, var, model, gender,
        c_status, hr, hr.low, hr.high
      )
    # output
    summary.models <- rbind(
      summary.models,
      working.summary
    )


    # cancers diagnosed 1-5y ago
    if (transition.name %in%
      c(
        "From general population to diagnosed with COVID-19",
        "From diagnosed with COVID-19 to hospitalised with COVID-19",
        "From general population to death"
      )) {
      m.working <- cph(as.formula(paste(
        "Surv(time, status)~ ",
        paste(
          cancers[i],
          "+gender+rcs(age,5)+
               medea+smoke.all_time+
               a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
               a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity"
        )
      )),
      surv = TRUE, x = TRUE, y = TRUE,
      data = transition.data %>%
        filter_se(paste0("cancer==0 |
                       (", cancers[i], "==1 & c_status_2cat==\"<5 years\")")),
      iter.max = 200
      )
    } else if (transition.name %in%
      c(
        "From general population to hospitalised with COVID-19",
        "From diagnosed with COVID-19 to death"
      )) {
      m.working <- cph(as.formula(paste(
        "Surv(time, status)~ ",
        paste(
          cancers[i],
          "+gender+pol(age,2)+
               medea+smoke.all_time+
               a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
               a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity"
        )
      )),
      surv = TRUE, x = TRUE, y = TRUE,
      data = transition.data %>%
        filter_se(paste0("cancer==0 |
                       (", cancers[i], "==1 & c_status_2cat==\"<5 years\")")),
      iter.max = 200
      )
    } else {
      m.working <- cph(as.formula(paste(
        "Surv(time, status)~ ",
        paste(
          cancers[i],
          "+gender+rcs(age,3)+
               medea+smoke.all_time+
               a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
               a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity"
        )
      )),
      surv = TRUE, x = TRUE, y = TRUE,
      data = transition.data %>%
        filter_se(paste0("cancer==0 |
                       (", cancers[i], "==1 & c_status_2cat==\"<5 years\")")),
      iter.max = 200
      )
    }
    working.summary <- as.data.frame(summary(m.working, antilog = FALSE))
    working.summary$var <- row.names(working.summary)
    working.summary$gender <- "Overall"
    working.summary$c_status <- "<5 years"
    working.summary$model <- "Fully adjusted"
    working.summary$transition.name <- transition.name
    working.summary <- working.summary %>%
      mutate(
        hr = exp(Effect),
        hr.low = exp(`Lower 0.95`),
        hr.high = exp(`Upper 0.95`)
      ) %>%
      select(
        transition.name, var, model, gender,
        c_status, hr, hr.low, hr.high
      )
    # output
    summary.models <- rbind(
      summary.models,
      working.summary
    )

    ## cancers diagnosed >5y ago
    if (transition.name %in%
      c(
        "From general population to diagnosed with COVID-19",
        "From diagnosed with COVID-19 to hospitalised with COVID-19",
        "From general population to death"
      )) {
      m.working <- cph(as.formula(paste(
        "Surv(time, status)~ ",
        paste(
          cancers[i],
          "+gender+rcs(age,5)+
                medea+smoke.all_time+
                a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
                a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity"
        )
      )),
      surv = TRUE, x = TRUE, y = TRUE,
      data = transition.data %>%
        filter_se(paste0("cancer==0 |
                       (", cancers[i], "==1 & c_status_2cat==\">5years\")")),
      iter.max = 200
      )
    } else if (transition.name %in%
      c(
        "From general population to hospitalised with COVID-19",
        "From diagnosed with COVID-19 to death"
      )) {
      m.working <- cph(as.formula(paste(
        "Surv(time, status)~ ",
        paste(
          cancers[i],
          "+gender+pol(age,2)+
                medea+smoke.all_time+
                a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
                a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity"
        )
      )),
      surv = TRUE, x = TRUE, y = TRUE,
      data = transition.data %>%
        filter_se(paste0("cancer==0 |
                       (", cancers[i], "==1 & c_status_2cat==\">5years\")")),
      iter.max = 200
      )
    } else {
      m.working <- cph(as.formula(paste(
        "Surv(time, status)~ ",
        paste(
          cancers[i],
          "+gender+rcs(age,3)+
                 medea+smoke.all_time+
                a_autoimmune_condition + a_chronic_kidney_disease + a_copd + a_dementia +
                a_heart_disease + a_hyperlipidemia + a_hypertension + a_t2_diabetes + a_obesity"
        )
      )),
      surv = TRUE, x = TRUE, y = TRUE,
      data = transition.data %>%
        filter_se(paste0("cancer==0 |
                       (", cancers[i], "==1 & c_status_2cat==\">5years\")")),
      iter.max = 200
      )
    }
    working.summary <- as.data.frame(summary(m.working, antilog = FALSE))
    working.summary$var <- row.names(working.summary)
    working.summary$gender <- "Overall"
    working.summary$c_status <- ">5years"
    working.summary$model <- "Fully adjusted"
    working.summary$transition.name <- transition.name
    working.summary <- working.summary %>%
      mutate(
        hr = exp(Effect),
        hr.low = exp(`Lower 0.95`),
        hr.high = exp(`Upper 0.95`)
      ) %>%
      select(
        transition.name, var, model, gender,
        c_status, hr, hr.low, hr.high
      )
    # output
    summary.models <- rbind(
      summary.models,
      working.summary
    )
  }
  summary.models
}
```



```{r}
# Gathering results by transition
summary.m.healthy.diagnosis <- get.models(
  cancers = cancers,
  transition.data = r.healthy.diagnosis,
  transition.name = "From general population to diagnosed with COVID-19"
)

summary.m.healthy.hospitalised <- get.models(
  cancers = cancers,
  transition.data = r.healthy.hospitalised,
  transition.name = "From general population to hospitalised with COVID-19"
)

summary.m.healthy.death <- get.models(
  cancers = cancers,
  transition.data = r.healthy.death,
  transition.name = "From general population to death"
)

summary.m.diagnosis.hospitalised <- get.models(
  cancers = cancers,
  transition.data = r.diagnosis.hospitalised,
  transition.name = "From diagnosed with COVID-19 to hospitalised with COVID-19"
)

summary.m.diagnosis.death <- get.models(
  cancers = cancers,
  transition.data = r.diagnosis.death,
  transition.name = "From diagnosed with COVID-19 to death"
)

summary.m.hospitalised.death <- get.models(
  cancers = cancers,
  transition.data = r.hospitalised.death,
  transition.name = "From hospitalised with COVID-19 to death"
)
```


# Summary of models
```{r}
estimates <- rbind(
  summary.m.healthy.diagnosis,
  summary.m.healthy.hospitalised,
  summary.m.diagnosis.hospitalised,
  summary.m.diagnosis.death,
  summary.m.hospitalised.death,
  summary.m.healthy.death
)

# save estimates
# save(estimates, file = "M:/Elena - R/Multi-state  cancer/results/3.estimates_years_solid_cancers.Rda")

# load("M:/Elena - R/Multi-state  cancer/results/3.estimates_years_solid_cancers.Rda")

# just the estimated effect of cancers
estimates <- estimates %>%
  filter(!(str_detect(var, "a_"))) %>%
  filter(var == "breast" | var == "prostate" |
    str_detect(var, paste(cancers, collapse = "|")))

summary.table <- estimates %>%
  mutate(est = paste0(
    nice.num2(hr), " (",
    nice.num2(hr.low), " to ",
    nice.num2(hr.high), ")"
  )) %>%
  mutate(group = paste0(model, "- ", gender, "- ", c_status)) %>%
  select(transition.name, var, group, est)

summary.table <- summary.table %>%
  pivot_wider(
    names_from = group,
    values_from = est
  )

# save results
# write.csv2(summary.table, "M:/Elena - R/Multi-state  cancer/results/3.HR_years_solid_cancers.csv", row.names = TRUE)
```

```{r}
# see results
kable(summary.table,
  col.names = c(
    "Transition", "Variable",
    "Overall", "Less than 5 years", "More than 5 years"
  )
) %>%
  add_header_above(c("", "", "", "Years since the diagnosis of cancer" = 2)) %>%
  kable_styling(bootstrap_options = c("striped", "bordered")) %>%
  collapse_rows(columns = 1)
```

#Plotting HR for each specific cancer type
```{r}
plot.data <- estimates %>%
  mutate(
    transition.name =
      ifelse(transition.name ==
        "From general population to diagnosed with COVID-19",
      "From general population\nto diagnosed\nwith COVID-19",
      ifelse(transition.name ==
        "From general population to hospitalised with COVID-19",
      "From general population\nto hospitalised\nwith COVID-19",
      ifelse(transition.name ==
        "From diagnosed with COVID-19 to hospitalised with COVID-19",
      "From diagnosed with\nCOVID-19 to hospitalised\nwith COVID-19",
      ifelse(transition.name ==
        "From diagnosed with COVID-19 to death",
      "From diagnosed with\nCOVID-19 to\ndeath",
      ifelse(transition.name ==
        "From hospitalised with COVID-19 to death",
      "From hospitalised\nwith COVID-19\nto death",
      ifelse(transition.name ==
        "From general population to death",
      "From general population\nto death",
      NA
      )
      )
      )
      )
      )
      )
  ) %>%
  mutate(transition.name = factor(transition.name,
    levels = c(
      "From general population\nto diagnosed\nwith COVID-19",
      "From general population\nto hospitalised\nwith COVID-19",
      "From diagnosed with\nCOVID-19 to hospitalised\nwith COVID-19",
      "From diagnosed with\nCOVID-19 to\ndeath",
      "From hospitalised\nwith COVID-19\nto death",
      "From general population\nto death"
    )
  ))

plot.data <- plot.data %>%
  mutate(var = str_to_title(var)) %>%
  mutate(var = ifelse(var == "Lung_trachea", "Lung", var)) %>%
  mutate(var = factor(var,
    levels = c(
      "Kidney",
      "Melanoma",
      "Lung",
      "Bladder",
      "Colorectal",
      "Prostate",
      "Breast"
    )
  )) %>%
  mutate(c_status = ifelse(c_status == "<5 years", "<5 years",
    ifelse(c_status == ">5years", "\u22655 years",
      c_status
    )
  )) %>%
  mutate(c_status = factor(c_status,
    levels = c("Overall", "\u22655 years", "<5 years")
  )) %>%
  mutate(hr_text = paste0(
    nice.num2(hr), " (",
    nice.num2(hr.low), " to ",
    nice.num2(hr.high), ")"
  ))
```


## Without the transition from general population to death
```{r, fig.width= 12}
# without from general population to death
hrs.fully.solid <- plot.data %>%
  filter(transition.name != "From general population\nto death") %>%
  mutate(c_status = relevel(c_status, "Overall")) %>%
  ggplot(aes(
    x = hr,
    y = var,
    xmin = hr.low,
    xmax = hr.high,
    colour = var
  )) +
  scale_x_log10(
    breaks = c(0.25, 0.5, 0.75, 1, 2, 5),
    labels = c("0.25", "0.5", "", "1", "2", "5"),
    limits = c(0.2, 6)
  ) +
  facet_grid(c_status ~ transition.name,
    drop = TRUE,
    switch = "y"
  ) +
  geom_point(size = 3, position = position_dodge(width = 1), shape = 20) +
  geom_errorbar(width = 0, size = 1, position = position_dodge(width = 1)) +
  scale_colour_manual(values = c("#6C6F75", "#CCAD00", "#2C08AF", "#0ABAFF", "#DE3CBB")) +
  geom_vline(
    xintercept = 1, colour = "#000000",
    linetype = 2
  ) +
  theme_bw() +
  theme(
    text = element_text(family = "A"),
    panel.spacing = unit(0, "lines"),
    legend.position = "bottom",
    legend.title = element_blank(),
    legend.text = element_text(size = 14),
    axis.title.y = element_blank(),
    axis.text.x = element_text(size = 10),
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    axis.title = element_text(size = 14),
    strip.text.x = element_text(size = 12, face = "bold"),
    strip.text.y.left = element_text(size = 14, angle = 0, face = "bold"),
    strip.background = element_rect(fill = "#f7f7f7")
  ) +
  xlab("\n aHR (with 95% CI)") +
  ylab("") +
  labs(title = "Estimated hazard ratios for cancer subtypes, overall and by years since cancer diagnosis
Reference: cancer-free patients. Fully adjusted models") +
  guides(shape = guide_legend(reverse = T), colour = guide_legend(reverse = T))

hrs.fully.solid

# ggsave("M:/Elena - R/Multi-state  cancer/results/gg.hrs.years.solid.cancers_cov.png", hrs.fully.solid,
#  dpi = 300,
#  width = 15, height = 9)
```



##Counts of each cancer per transition / years since cancer diagnosis

```{r}
vars <- c(
  "hematological",
    "solid",
  "breast",
  "prostate",
  "colorectal",
  "bladder",
  "lung_trachea",
  "kidney",
  "melanoma"
  )


factor.vars <- c(
  "hematological",
    "solid",
  "breast",
  "prostate",
  "colorectal",
  "bladder",
  "lung_trachea",
  "kidney",
  "melanoma"
  )

library(tableone)
library(kableExtra)

# counts by transition, events and cancer types
# healthy to diagnosis
counts.healthy.dx <- print(CreateTableOne(
  vars = vars,
  factorVars = factor.vars,
  includeNA = T,
  data = r.healthy.diagnosis,
  strata = c("status"),
  test = F
),
showAllLevels = F, smd = F,
nonnormal = vars,
noSpaces = TRUE,
contDigits = 1,
format = "f", # include only counts
printToggle = FALSE
)

colnames(counts.healthy.dx) <- c("Healthy_no_event", "Healthy_dx")
counts.healthy.dx <- counts.healthy.dx[,-1]

# healthy to hospitalised
counts.healthy.hx <- print(CreateTableOne(
  vars = vars,
  factorVars = factor.vars,
  includeNA = T,
  data = r.healthy.hospitalised,
  strata = c("status"),
  test = F
),
showAllLevels = F, smd = F,
nonnormal = vars,
noSpaces = TRUE,
contDigits = 1,
format = "f", # include only counts
printToggle = FALSE
)

colnames(counts.healthy.hx) <- c("Healthy_no_event", "Healthy_hx")
counts.healthy.hx <- counts.healthy.hx[,-1]

# dianosis to hospitalised
counts.dx.hx <- print(CreateTableOne(
  vars = vars,
  factorVars = factor.vars,
  includeNA = T,
  data = r.diagnosis.hospitalised,
  strata = c("status"),
  test = F
),
showAllLevels = F, smd = F,
nonnormal = vars,
noSpaces = TRUE,
contDigits = 1,
format = "f", # include only counts
printToggle = FALSE
)
colnames(counts.dx.hx) <- c("Diag_no_event", "Diag_hospi")

counts.dx.hx <- counts.dx.hx[,-1]

# dianosis to death
counts.dx.death <- print(CreateTableOne(
  vars = vars,
  factorVars = factor.vars,
  includeNA = T,
  data = r.diagnosis.death,
  strata = c("status"),
  test = F
),
showAllLevels = F, smd = F,
nonnormal = vars,
noSpaces = TRUE,
contDigits = 1,
format = "f", # include only counts
printToggle = FALSE
)
colnames(counts.dx.death) <- c("Diag_no_event", "Diag_Death")

counts.dx.death <- counts.dx.death [,-1]

# hospitalised to death
counts.hx.death <- print(CreateTableOne(
  vars = vars,
  factorVars = factor.vars,
  includeNA = T,
  data = r.hospitalised.death,
  strata = c("status"),
  test = F
),
showAllLevels = F, smd = F,
nonnormal = vars,
noSpaces = TRUE,
contDigits = 1,
format = "f", # include only counts
printToggle = FALSE
)
colnames(counts.hx.death) <- c("Hosp_no_event", "Hosp_Death")
counts.hx.death <- counts.hx.death[,-1]


counts.all <- cbind(
  counts.healthy.dx,
  counts.healthy.hx,
  counts.dx.hx,
  counts.dx.death,
  counts.hx.death
)


## <5 year
# healthy to diagnosis
counts.healthy.dx <- print(CreateTableOne(
  vars = vars,
  factorVars = factor.vars,
  includeNA = T,
  data = r.healthy.diagnosis%>% filter(c_status_2cat=="<5 years"),
  strata = c("status"),
  test = F
),
showAllLevels = F, smd = F,
nonnormal = vars,
noSpaces = TRUE,
contDigits = 1,
format = "f", # include only counts
printToggle = FALSE
)

colnames(counts.healthy.dx) <- c("Healthy_no_event", "Healthy_dx")
counts.healthy.dx <- counts.healthy.dx[,-1]

# healthy to hospitalised
counts.healthy.hx <- print(CreateTableOne(
  vars = vars,
  factorVars = factor.vars,
  includeNA = T,
  data = r.healthy.hospitalised %>% filter(c_status_2cat=="<5 years"),
  strata = c("status"),
  test = F
),
showAllLevels = F, smd = F,
nonnormal = vars,
noSpaces = TRUE,
contDigits = 1,
format = "f", # include only counts
printToggle = FALSE
)

colnames(counts.healthy.hx) <- c("Healthy_no_event", "Healthy_hx")
counts.healthy.hx <- counts.healthy.hx[,-1]

# dianosis to hospitalised
counts.dx.hx <- print(CreateTableOne(
  vars = vars,
  factorVars = factor.vars,
  includeNA = T,
  data = r.diagnosis.hospitalised%>% filter(c_status_2cat =="<5 years"),
  strata = c("status"),
  test = F
),
showAllLevels = F, smd = F,
nonnormal = vars,
noSpaces = TRUE,
contDigits = 1,
format = "f", # include only counts
printToggle = FALSE
)
colnames(counts.dx.hx) <- c("Diag_no_event", "Diag_hospi")

counts.dx.hx <- counts.dx.hx[,-1]

# dianosis to death
counts.dx.death <- print(CreateTableOne(
  vars = vars,
  factorVars = factor.vars,
  includeNA = T,
  data = r.diagnosis.death%>% filter(c_status_2cat =="<5 years"),
  strata = c("status"),
  test = F
),
showAllLevels = F, smd = F,
nonnormal = vars,
noSpaces = TRUE,
contDigits = 1,
format = "f", # include only counts
printToggle = FALSE
)
colnames(counts.dx.death) <- c("Diag_no_event", "Diag_Death")

counts.dx.death <- counts.dx.death [,-1]

# hospitalised to death
counts.hx.death <- print(CreateTableOne(
  vars = vars,
  factorVars = factor.vars,
  includeNA = T,
  data = r.hospitalised.death%>% filter(c_status_2cat =="<5 years"),
  strata = c("status"),
  test = F
),
showAllLevels = F, smd = F,
nonnormal = vars,
noSpaces = TRUE,
contDigits = 1,
format = "f", # include only counts
printToggle = FALSE
)
colnames(counts.hx.death) <- c("Hosp_no_event", "Hosp_Death")
counts.hx.death <- counts.hx.death[,-1]


counts.5y <- cbind(
  counts.healthy.dx,
  counts.healthy.hx,
  counts.dx.hx,
  counts.dx.death,
  counts.hx.death
)


colnames(counts.5y) <- str_replace(colnames(counts.5y), "counts", "5y")
colnames(counts.all) <- str_replace(colnames(counts.all), "counts", "All")
counts <- cbind(counts.all, counts.5y)
rownames(counts) <- str_replace(rownames(counts), " = 1", "")

counts <- counts[-1,]
rownames(counts) <- str_to_sentence(rownames(counts))


```



```{r}

kable(counts,
      col.names = c(
        "To diagnosed with COVID-19",
        "To hospitalised with COVID-19",
        "To hospitalised with COVID-19",
        "To death",
        "To death",
        "To diagnosed with COVID-19",
        "To hospitalised with COVID-19",
        "To hospitalised with COVID-19",
        "To death",
        "To death"
      )
) %>%
  add_header_above(c(
    " " = 1,
    "From general population" = 2,
    "From diagnosed with COVID-19" = 2,
    "From hospitalised with COVID-19" = 1,
    "From general population" = 2,
    "From diagnosed with COVID-19" = 2,
    "From hospitalised with COVID-19" = 1
  )) %>%
  add_header_above(c(
    " " = 1,
    "Overall" = 5,
    "<5 year cancer patients" = 5
  )) %>%
  kable_styling(bootstrap_options = c("striped", "bordered"))
```
